Move RealTorServiceConfig to internal file prefixed with -
