+++ library/manager/kmp-tor-manager/src/jvmTest/kotlin/io/matthewnelson/kmp/tor/manager/internal/actions/ActionQueueUnitTest.kt
@@ -35,7 +35,7 @@ class ActionQueueUnitTest {
 
     private val lock = Mutex()
     private var lockCount = 0
-    private val queue = ActionQueue() // RealActionQueue
+    private val queue = ActionQueue.newInstance() // RealActionQueue
     private var dispatcher: CoroutineDispatcher? = null
 
     @Before
